Change files-match-result to urls-match-result.
* Uses std::fs::read_to_string to skip a bunch of file and I/O trait pain.
