+++ exercise-book/src/SUMMARY.md
@@ -11,7 +11,7 @@
 - [SimpleDB](./simple-db.md)
   - [Knowledge](./simple-db-knowledge.md)
   - [Step-by-Step Solution](./simple-db-solution.md)
-- [Files, match, result](./files-match-result.md)
+- [URLs, match, result](./urls-match-result.md)
 
 # Applied Rust
