--- exercise-solutions/files-match-result/src/main.rs
@@ -1,6 +1,0 @@
-fn main() {
-    fn main() {
-        println!("This is the solution for the files-match-result exercise.");
-        println!("Run like 'cargo run --example step_3' to run the Step 3 solution");
-    }
-}
