+++ exercise-solutions/urls-match-result/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["Ferrous Systems"]
 edition = "2021"
-name = "files_match"
+name = "urls_match_result"
 version = "0.1.0"
 
 [dependencies]
