+++ exercise-solutions/urls-match-result/src/main.rs
@@ -1,0 +1,4 @@
+fn main() {
+    println!("This is the solution for the urls-match-result exercise.");
+    println!("Run like 'cargo run --example step_3' to run the Step 3 solution");
+}
